+++ artiq/coredevice/comm_kernel.py
@@ -1,5 +1,7 @@
 import struct
 import logging
+import socket
+import sys
 import traceback
 import numpy
 from enum import Enum
@@ -69,29 +71,60 @@ class RPCReturnValueError(ValueError):
 RPCKeyword = namedtuple('RPCKeyword', ['name', 'value'])
 
 
-class CommGeneric:
-    def __init__(self):
+def set_keepalive(sock, after_idle, interval, max_fails):
+    if sys.platform.startswith("linux"):
+        sock.setsockopt(socket.SOL_SOCKET, socket.SO_KEEPALIVE, 1)
+        sock.setsockopt(socket.IPPROTO_TCP, socket.TCP_KEEPIDLE, after_idle)
+        sock.setsockopt(socket.IPPROTO_TCP, socket.TCP_KEEPINTVL, interval)
+        sock.setsockopt(socket.IPPROTO_TCP, socket.TCP_KEEPCNT, max_fails)
+    elif sys.platform.startswith("win") or sys.platform.startswith("cygwin"):
+        # setting max_fails is not supported, typically ends up being 5 or 10
+        # depending on Windows version
+        sock.ioctl(socket.SIO_KEEPALIVE_VALS,
+                   (1, after_idle*1000, interval*1000))
+    else:
+        logger.warning("TCP keepalive not supported on platform '%s', ignored",
+                       sys.platform)
+
+
+def initialize_connection(host, port):
+    sock = socket.create_connection((host, port), 5.0)
+    sock.settimeout(None)
+    set_keepalive(sock, 3, 2, 3)
+    logger.debug("connected to host %s on port %d", host, port)
+    return sock
+
+
+class CommKernel:
+    def __init__(self, dmgr, host, port=1381):
         self._read_type = None
+        self.host = host
+        self.port = port
 
     def open(self):
-        """Opens the communication channel.
-        Must do nothing if already opened."""
-        raise NotImplementedError
+        if hasattr(self, "socket"):
+            return
+        self.socket = initialize_connection(self.host, self.port)
+        self.socket.sendall(b"ARTIQ coredev\n")
 
     def close(self):
-        """Closes the communication channel.
-        Must do nothing if already closed."""
-        raise NotImplementedError
+        if not hasattr(self, "socket"):
+            return
+        self.socket.close()
+        del self.socket
+        logger.debug("disconnected")
 
     def read(self, length):
-        """Reads exactly length bytes from the communication channel.
-        The channel is assumed to be opened."""
-        raise NotImplementedError
+        r = bytes()
+        while len(r) < length:
+            rn = self.socket.recv(min(8192, length - len(r)))
+            if not rn:
+                raise ConnectionResetError("Connection closed")
+            r += rn
+        return r
 
     def write(self, data):
-        """Writes exactly length bytes to the communication channel.
-        The channel is assumed to be opened."""
-        raise NotImplementedError
+        self.socket.sendall(data)
 
     #
     # Reader interface